+++ library-auth/src/main/java/com/onetuks/libraryauth/jwt/filter/JwtAuthenticationFilter.java
@@ -1,5 +1,8 @@
-package com.onetuks.libraryauth.jwt;
+package com.onetuks.libraryauth.jwt.filter;
 
+import com.onetuks.libraryauth.jwt.service.model.AuthToken;
+import com.onetuks.libraryauth.jwt.service.AuthTokenProvider;
+import com.onetuks.libraryauth.jwt.util.AuthHeaderUtil;
 import com.onetuks.libraryauth.service.AuthService;
 import jakarta.annotation.Nonnull;
 import jakarta.servlet.FilterChain;
