+++ library-auth/src/main/java/com/onetuks/libraryauth/jwt/service/AuthTokenService.java
@@ -1,0 +1,8 @@
+package com.onetuks.libraryauth.jwt.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class AuthTokenService {
+
+}
