+++ library-auth/src/main/java/com/onetuks/libraryauth/oauth/strategy/ClientProviderStrategy.java
@@ -1,6 +1,6 @@
 package com.onetuks.libraryauth.oauth.strategy;
 
-import com.onetuks.libraryauth.oauth.dto.ClientAuthToken;
+import com.onetuks.libraryauth.oauth.strategy.dto.auth_token.ClientAuthToken;
 import com.onetuks.librarydomain.member.model.vo.AuthInfo;
 
 public interface ClientProviderStrategy {
