+++ library-auth/src/main/java/com/onetuks/libraryauth/oauth/strategy/dto/auth_token/KakaoAuthToken.java
@@ -1,4 +1,4 @@
-package com.onetuks.libraryauth.oauth.dto;
+package com.onetuks.libraryauth.oauth.strategy.dto.auth_token;
 
 public record KakaoAuthToken(
     String token_type,
