+++ library-auth/src/main/java/com/onetuks/libraryauth/oauth/strategy/dto/auth_token/NaverAuthToken.java
@@ -1,4 +1,4 @@
-package com.onetuks.libraryauth.oauth.dto;
+package com.onetuks.libraryauth.oauth.strategy.dto.auth_token;
 
 public record NaverAuthToken(
     String access_token,
