+++ library-auth/src/main/java/com/onetuks/libraryauth/oauth/strategy/dto/user_info/GoogleUserInfo.java
@@ -1,4 +1,4 @@
-package com.onetuks.libraryauth.oauth.dto;
+package com.onetuks.libraryauth.oauth.strategy.dto.user_info;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -9,13 +9,13 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @JsonNaming(SnakeCaseStrategy.class)
-public class GoogleUser {
+public class GoogleUserInfo {
 
   private String sub;
   private String email;
   private String name;
 
-  public GoogleUser(String sub, String email, String name) {
+  public GoogleUserInfo(String sub, String email, String name) {
     this.sub = sub;
     this.email = email;
     this.name = name;
